+++ src/main/java/io/github/teampropulsive/DataGeneration.java
@@ -1,4 +1,4 @@
-package io.github.teampropulsive.util;
+package io.github.teampropulsive;
 
 import io.github.teampropulsive.util.AdvancementProvider;
 import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
